+++ lib/src/banner/options.dart
@@ -1,5 +1,8 @@
 class BannerAdOptions {
+  /// Reload the ad whenever the orientation changes. Defaults to `true`
   final bool reloadWhenOrientationChanges;
+
+  /// Reload the ad whenever its size changes. Defaults to `true`
   final bool reloadWhenSizeChanges;
 
   const BannerAdOptions({